fix: rework swapper loader to be based on RTK state (#10215)
* fix: rework swapper loader to be based on RTK state

---------
